feat(view): stash all those $$ in da view
